refactor: The initial purpose of the validation class was to be used for web user API. However, there wasn't as much overlap with validation used for AdminInvitesUserForm. So move validation for Web User Resource its own file and create a class to organize the validations relevant to AdminInvitesUserForm
